Create UserCanChangeEmail Hook
There is PrefsEmailAuditHook and its documentation
said it can be used to abort the changing, but the
return value of the call to the hook is not used
anywhere.

But making the return value of PrefsEmailAuditHook
be used is limited because the hook is triggered
after a confirmation email is sent, and the name
maybe hints its main usage is audit.

So, creates a new hook to give the opportunity to
abort the changing to extensions.

Bug: T315128
Change-Id: I6be98f018af2c7334efb21a423b2092f82941c3b