EncryptedPassword: Store default parameters as strings
Currently ->needsUpdate always returns true since it compares 0 as an
integer to "0" as a string strictly, which fails.

This means that on every successful login the user password hash is
newly generated and stored in the db, causing unnecessary writes to the
user table.

Also cast params['secret'] to int since that is read from the stored
password hash and thus will be a string and is already being implicitly
casted.

Bug: T366130
Change-Id: Icad0dcf15d5889f8cc7941a7ea7bf250ee30f47c
(cherry picked from commit 1bfde4aef92533f6bb0a7862bbe8ed687d2c4af0)