Get mobile url for Parsoid's baseHref
The legacy parser does not run ExpandToAbsoluteUrls unless it's doing
?action=render.  ExpandToAbsoluteUrls doesn't work for mobile urls,
which seems to be captured in T171398 / T195494.  Since relative urls
aren't resolved in legacy output though, the browser uses the mobile
url.

Parsoid, however, does ExtractBody which has its own expandRelativeAttrs
pass, which resolves relative urls against the baseHref in the document
head.  The baseHref is taken from MainConfigNames::Server, which
presumably suffers the same issue as the above task.  But also maybe MFE
is transforming cached html, where the non-mobile baseHref is desirable.

In any case, to produce the same urls as the legacy parser, transform
the baseHref to one that conforms with mobile url template.

Bug: T365483
Change-Id: I32800f5ea848d70b6ef67ec9102c432b9626afcb