Fix bundle reinjection of ContentDOMTransformStage
When re-injecting the page bundle to the newly created ParserOutput, we
were omitting the version, headers and contentmodel data of said page
bundle reinjection. This patch fixes that.

Note that it will silence places where getText should typically not be
called, but that's a larger problem that needs to be addressed on the
calling places, and doesn't detract from the fact that we needed to fix
this loss of information on the bundle anyway.

Bug: T365433
Depends-On: I2a87a8233b9e42cbafdba63bdf513abe00d826ce
Change-Id: I7f57ddc76b9d3b24226f8b5da1b70bc83134856f