Don't ignore offsetType attribute on lint API paths
The default used to be 'ucs2' when linting but
Ic9b7cc0fcf365e772b7d080d76a065e3fd585f80 stopped setting the offsetType
in the environment options, which changed the output to Parsoid's
default, 'byte'.

For the parsing paths, pagebundles are now post-processed and
non-pagebundles were left as a TODO.

However, for the lint paths, it was always ok to continue setting the
environment option because Parsoid is called directly and doesn't go
through the ParserOutputAccess.  The above patch was trying to limit
Parsoid specific options to that.

Bug: T365284
Change-Id: I8389c2f53b399b39a9f1d908a38aecb3abcb15ef