mediaHandler: Respect image scaler configs for Webp and XCF files
ImageMagick should not be used if it's not installed. When a custom
convert command is configured, it's the site admin's responsibility
to make sure it works with Webp and XCF files. The php-imagick
extension is just a wrapper for the ImageMagick library, and would
also work for Webp and XCF files.

Bug: T308386
Change-Id: I9fcf22b0c4b6d5d132e5a44530e3f255075f7fb4