Fix setting offsetType in options passed to HtmlInputTransformHelper
Ic9b7cc0fcf365e772b7d080d76a065e3fd585f80 stopped setting it as an
envOptions so pull it from the attribs when constructing.

The tests continued to set the offsetType in the envOptions, despite the
ParsoidHandler no longer doing that, so they failed to catch the error.

HtmlInputTransformHelper still considers it an envOptions though and a
cleanup could happen there too.  Although, offsetType still is an
envOptions to Parsoid, so maybe it's fine as is.

Change-Id: I661a170fbb2c61308e32a43a9f6b0ce8d2ffb797