Maintenance: Consistently print a newline after errors
This code path used to add a newline too until it was removed in
bb45c5e8a3034a24fa7c85f27c4ce035de56ccb0 (2011). It seems to be an
accidental change, since other changes in that commit carefully kept
the newlines. I guess it's not reached often in practice…

Change-Id: I1f0e9735efed70fc50e6b4592a7b643f6f51ddc6