Optimize globals processing in ExtensionRegistry for performance
We discussed this in the previous patch Ic6b8841 but wanted to keep
it focussed on one change at a time.

This code path can be quite hot the more extensions are registered.
This slightly re-arranges the existing code to avoid various kinds of
duplicate calls. No behavior changes.

I tried hard to not compromise readability for performance. In fact
I think this is even better readable.

Bug: T366547
Change-Id: I3c1263e0c0ed78bac1044d0287830cb1a32b7e44