HTMLForm: Remove use of jquery.chosen in HTMLMultiSelectField
HTMLMultiSelectField used the jquery.chosen library, but only for the
legacy display format, and only if the 'dropdown' option was passed.
There is only one form that passes the 'dropdown' option
(SpecialActiveUsers), and that form uses the OOUI display format, so
this code is unused.

Also remove the htmlform-chosen-placeholder i18n message, which is now
unused as well.

Bug: T53443
Change-Id: Ie4f9fff322ae26fbdf9eae1b3634d1b04e746179