Remove unnecessary try/catch in ParsoidCachePrewarmJob
The two exceptions caught here should never be thrown in actual practice,
and indicate programming errors if they are thrown: the first would only
occur for an invalid language variant code or invalid DOM name characters,
and the second would only be thrown if the wikitext size limit is exceeded,
which should never happen with articles retrieved from the database.

(Parsoid's handling of overlarge articles also differs from legacy,
and it should probably match legacy and just truncate to the size
limit rather than throw an exception, but that's a separate issue.)

In any case, there's not much gained here by catching and rethrowing
rather than just letting the original Parsoid exception, were it to
ever occur, bubble up and be logged directly as a uncaught exception.

Followup-To: I96161a64952e1809c0aec773d5a3dd4c71105657
Change-Id: Ibac2d01306d0637097c0cc3363e63f238bb649fe