Fix test about applying mw-selflink-fragment to variant links
A byte sequence (d0 b0) was presumably erroneously introduced in the
link [[Duna#Foo]] which prevented it from being recognized as the same
target as the title associated with the current test.  An article (Дуна)
that's a variant of Duna exists, and variant conversion determines that
the link points at that page instead.  Removing those bytes updates the
output to add the mw-selflink-fragment class.

The test itself is updated to reflect that, as of 1fca3b5b, fragment
links are selflinks and a FIXME is added for a follow-up to correct the
variant fragment links to add the class.

Change-Id: Iee0326f85c919a672397d0378f3549f583c17e28