ApiQuerySiteinfo: Support $wgGroupInheritsPermissions
I noticed that the "confirmed" group was missing from the API result
on en.wikipedia.org. It is defined using $wgGroupInheritsPermissions,
not $wgGroupPermissions.

Rather than only using the latter, obtain lists of user groups and
granted permissions from UserGroupManager and GroupPermissionsLookup
respectively (and not just for filtering undefined groups out of the
lists of changeable groups).

Bug: T357846
Change-Id: I473868449dcf360afcba3c731ac7f282b623b1cd