Initialize calendar with default date in DateTimeInputWidget
When initializing the field, also initialize the calendar so that the
chosen date shows as selected. This uses the same logic used by the
onChange handler.

I believe it would be possible to set the `selected` property in the
calendar configuration instead, but since callers can override the
calendar configuration, that is not done here.

Bug: T367999
Change-Id: I7dae6e6ad1f3f3ae200a2e80ad051fa404eb4f55