Ensure noresize also works when placed on figure
In mobile, noresize only has to be a parent of an img to take effect.
In these newer styles however, we scope to figure as we do not want to
influence inline images.

It is possible to use File syntax's class= to set noresize on figure.
Additionally the Imagemap extension places noresize on the figure element

Bug: T368034
Change-Id: I15f550f7be31c22baef5267708697310c7d9f8a4