rdbms: Reduce log severity of "found writes pending"
What:
Reduce log severity of "found writes pending" message from warning to
info.

Why:
During issues and slow downs, this triggers a massive flood of logs.
For example, in the outage of this weekend (connected to the ticket) it
triggered 40M logs and 40Gbit/s network flow which brought everything
down.
I'm not seeing any reason this should be a warning, there is no chance
of data corruption or major issues or partial commits (from my understanding).
I might be wrong though.

Bug: T368289
Change-Id: I7b83271858101f10f39cff4da539d5d653507d5a