Modify WikiExporter's BATCH_SIZE from 50000 to 10000
On T207628, as part of performance tuning, WikiExporter's BATCH_SIZE
was set to 50000. There was no decisive benchmarking done to support
that size. We have recently seen a lot of database pressure coming from
this configuration.

We now set it to 10000. The thinking is that this may not affect the
runtime of the dumps in a significant way, while alleviating strain
on the databases.

Bug: T368098
Change-Id: I2fd67c656c5c57fd94165bab22a62edae8f5f46a
(cherry picked from commit e97a52e32770c4bf8c911fc53cdcbcde62ba86c3)