i18n: Add messages for Codex components
Some Codex components have strings that are always or usually
consistent across features (such as the aria-label for the
icon-only "Close" button in a CdxDialog). Currently, the burden
is on the developer user to provide a translated string every
time they use one of these components.

These new messages will be used in Codex components for features
in MediaWiki, so the strings no longer need to be provided by the
dev user. See the parent task T345386 for additional context.

Bug: T368447
Change-Id: Ia39b9b775cc53719deae51abe96e18ba13dcfb7d