tests: Support temp accounts in MockAuthorityTrait
With temp accounts becoming a new kind of user besides logged-out and
named users, it makes sense to provide utility methods for testing
permissions with them. This is especially true as 1) more and more code
becomes aware of temp users, and 2) Authority itself has a `isTemp`
method, but its return value cannot be altered with the existing methods
in this trait.

Also change some documented union types to intersection types where that
makes sense.

Change-Id: I83f079bf3c39843adb5279ab325ea1fc2a127609