OutputPage: Convert helplink icon to use codex css mixin
Migrate this icon to codex for the purposes of having it appear in night
mode correctly.  Additionally, clean up the svg as it is no longer used

A couple notes:
1. This is not entirely visually identical to the current method - in
practice it ends up being nearly imperceptible, but worth noting that
even with the padding there is a slight vertical shift if you compare
them side by side
2. Due to the approach taken, I'm making the intentional choice to not
make this backwards compatible with cached HTML.  This means a cached
page will not have an icon for the duration it is cached, which feels
like an acceptable intermediary state, but feel free to disagree with me
in code review!

Slight visual change due to aforementioned color and position difference

Bug: T366358
Change-Id: I1c53e1c501242f9f9ab88ef8e0e4e85fb1971fbe