htmlform: Improve documentation of the `validation-callback` option
The code works with StatusValue, not just Status, so update the docblock
accordingly, especially considering the fact that more modern code
should prefer the former.

Also mention what the expected signature is by linking directly to the
callback property, so that developers don't have to scan the whole body
of `validate()` to figure it out.

Change-Id: I9d5f155d1d797e3fa8ab4a0f10cc19b21d933c8a