specials: Use username from query on Special:ListFiles
The query includes the username of the uploader via actor_name, use this
name for the LinkBatch instead of an extra query to resolve the userid
to username via UserCache.
The userid query was needed before the actor migration to ensure renamed
user are loaded/displayed correctly (UserCache was used before
a3b7ea6732 to display the username).

Include the file name into the batch to improve performance,
LinkRenderer::makeKnownLink is used, but for the redirect check a query
is still needed, but get now preloaded via the LinkBatch.

Change-Id: I73556c8449dfef38ba579c7dedf308a1d8af68fc