MediaWikiIntegrationTestCase: Cache needsDB()
This method is called relatively often (especially now that $this->db is
being replaced with $this->getDb() everywhere, T316841), and it’s not
entirely cheap: the underlying isTestInDatabaseGroup() method needs to
create a ReflectionClass for the test class and run a regex on its doc
comment. As the result is a constant for any single class (its doc
comment cannot change), we can easily cache it.

In a simple benchmark, caching the result reduces the time for 10
million self::needsDB() calls from 7.97 seconds to 2.28 seconds, and for
10 million $this->getDb() calls from 68.70 seconds to 61.60 seconds –
not a huge difference, but not nothing either. Some logging in changes
Ia7f622ff87 and I4454c00958 indicates that there are about a dozen test
classes (in the extensions pulled in by MediaWiki core and/or Wikibase)
that call needsDB() at least a thousand times, and two test classes
(ResourcesTest and ScopeStructureTest) that call it at least ten
thousand times; based on this, I think the expected savings are
somewhere between a tenth and a hundredth of a second per test run.

The “indexed by class name” trick to get per-class static method
variables is directly taken from the PHP RFC [1]; prior to PHP 8.1, a
plain static boolean would have worked as we need it in this case.

[1]: https://wiki.php.net/rfc/static_variable_inheritance#backward_incompatible_changes

Change-Id: I7917cb58aa0c86428798a207f7ad63484e8eb8ad