The "new" class should be added to the link as well as the LI element
Historically the new class has been added to the list element item
instead of the link but this is inconsistent with how the new
link is applied in article content. I recommend we add the new
class to both LI and A element going forward. On long term we should
aim to only add it to A element.

This allows for more obvious rendering without needing to resort
to special CSS to override certain cases.

Bug: T367982
Change-Id: I9ab0b1b2de2016edabb53bfcdd3f55bb8d687ccb