Special:AllMessages: Make colors compatible with dark mode
This uses a yellowish color for the default messages, similar to what
was before but a blueish color instead of previous green color. Maybe
we could use success color but that semantically was less related and
couldn't be paired with something related either but this way this is
semantically more related to diff capabilities of MediaWiki also.

Bug: T367827
Change-Id: I170b1a794713efe22dc38a69c3665ac01b3dbf76