specials: Add user tool links to uploader on Special:ListFiles
The uploader on Special:ListFiles just links to the user page,
also add a link to the talk page and to Special:Contributions as happen
on many other special pages and the file history on the file page
itself.

This also shows the uploader name, if ips are allowed to upload files.

Change-Id: Ia9f7ef6056ce3203bda5ff4af2ac45d09438e6af