api: Use much more specific PHPDoc type hints in ApiPageSet
A lot of arrays cleverly used as maps and sets in here. I believe
it's very much worth documenting them all as precise as possible.
This does have potential ripple effects on all code that uses these.

Naturally this can't have an effect in production as it's
exclusively about PHPDoc documentation.

Change-Id: I454baef2d871cbbaf0bb85886774cf74728b692a