[Figures] Remove redundant background color interfering with invert
Like Wickie37 pointed out in the following comment:
https://phabricator.wikimedia.org/T367589#9894432

The issue here is that the background color is being redefined at the
level of the file element, which is causing it to be inverted by the
skin-invert-image rule.  Since the parent element also has the same
background color defined, we can remove this, as it seems it was added
because these elements originally did not have the same color according
to I96e073234eb62ad968adb7f87e9c9d1082febb3c

Feel free to correct me if there's historical context I'm lacking!

Bug: T367589
Change-Id: I0abbfffb9fb7ff331a29e302c2afe1dd7d8a1b9a