objectcache: Remove unused key (`$i`) in `foreach()` loop
In I6b3167cac824d8bd8773bc66c386f41e4d380021 (75aec3686a2e3f4d0),
awesome Krinkle did some huge refactor of the BagOStuff parent and
subclasses to reduce boiler plate and indirection logic around the
`makeKey()` method.

In the process of doing that, the key that was previously used in
the foreach() loop in the `makeKeyInternal()` method was no longer
used.

The main intention of the `$i` was to append a colon(:) delimiter
infront of the keyspace before concatenating the various components
values to construct a cache key for the given keyspace and components.

Now that it's been handled differently, it's time to say farewell
to the `$i`.

Change-Id: I03a4b85509a9b841b05d000afb6242f91e06d9be