Add all necessary languages to line-height fixing
The headings line-height must be fixed for all
the languages of South and Southeast Asia that
are written in Brahmic scripts. Several of them
were missing, and they are added in this patch.

(It also puts a few of them that were in wrong places
in the correct alphabetical order.)

Change-Id: I4cecc2771c375bdf1912fbd4c3c1badc75af1d39