api: Fix incomplete docs for ApiPageSet::$mRequestedPageFields
Unfortunately Phan gets super confused here, which is why I had to
split this off from I454baef. While it might be possible to silence
Phan I think the better solution is to get rid of the problematic
null initialization. The null is never used for anything, anywhere.
The contrary. In ApiPageSet::processDbRow() the null values are
accessed with an array operator as if they are arrays, which
(currently) forces PHP to silently initialize it to actually be
an empty array. Instead I suggest to make this magic behavior
explicitely visible.

Change-Id: If7c4f49e46cc6050942611a7b2bca46ea550b31c