ResourceLoader: Inject installed skins to ResourceLoader Context
RL\Context::class depends on SkinFactory to get the installed skins
to check if the skin from the URL is in the list of installed skins
or defaults to 'fallback'.

Most instantiators of RL\Context already have a fallback skin specified
so instead of relying on SkinFactory in the constructor of Context,
let's get the list of installed skins in RL entry point and pass that
instead.

In tests, OutputPage, etc, they're all using a fake request which will
eventually default to the 'fallback' or 'apioutput' or related skin.

NOTE: In ContextTest.php, adjust callers of ::getResourceLoader() to
use `self` instead of `$this` since it's defined as a static method.

Bug: T32956
Change-Id: I0191a812f044a5c04c50c62163395c4a706c375c