MoveLogFormatter::getPreloadTitles: Handle bad titles
If the namespace configuration changed, it’s possible that the title in
the params can no longer be parsed. In that case, we shouldn’t return a
null entry, but just an empty array instead. (I haven’t tried it out
locally, but the test suggests that the actual formatting of the row
works without issue, with the title being shown as Special:BadTitle.)

Bug: T370396
Change-Id: I6335e12efa518332f7cd2268a833f8b60b8e6abe