REST: showcase usage of ArrayDef convenience functions
This demonstrates how to use ArrayDef::makeObjectSchema to define a
schema for a complex body field.

The "latest" field in the request body for page PUT requests identifies
the base revision. To allow the base revision info to be looped through
from an earlier GET request, it's nice to allow the "timestamp" field
to be present, even though we don't use it.

I changed the relevant part of Update.js to test that use case specifically.

Bug: T368131
Change-Id: I166396e0dbfc995e5346252ee438c9afe2c808e2