ResourceLoader: Hardcode the default math before Less.php 5.0 upgrade
v5.0 of less.php defaults to only doing division in parenthesis. From this task T368921, we see this will change the behaviour of a lot of extensions. Pending when that task is completed, we could set the default in ResourceLoader to always do division in or outside  parenthesis.

This is separate from I8cfb6aa11ac6766563221023c0, because otherwise
the mediawiki/vendor patch (Id70b0f4188e70f3e4b37a93a0) won't pass CI.

Bug: T368921
Change-Id: Ia23138114030b44e72ecd89edc73ee565f1fba39