ResourceLoader: Cache the state of special RL modules
In 80e5b160e098530 (I605b8cd1e1fc009b466), awesome @Krinkle did some
refactoring in OutputPage and introduced RL ClientHtml.php.

This class has a method (`getData()`) for computing the state of
special RL modules and a class property `$this->data` whose intention
I think was to cache the result of this computation for reuse by diff
callers (getHeadHtml(), getBodyHtml() etc).

But as it is today, the member is unused and this patch does a small
optimization by caching the result into the member and reusing if needed.

Bug: T370451
Change-Id: I73f21caace9d35bc71955ab147b02eb604c42948