tests: Remove unstyled text atop Special:JavaScriptTest
I added this over 10 years ago when this page was still an integrated
and skinned MW page. Now that we use the standalone QUnit layout, the
message no longer fits in. Also, since QUnit 2.14, the layout is fixed
so our message is invisible anyway becaise it falls behind the toolbar.

Bug: T250045
Change-Id: Icc228e332155147c7f88fb3d85953fc1046473fc