Stats: `copyToStatsdAt()` should not treat gauges as counters
In I15d6f100a31fbb98ad (13f556ab3ab4c50e63), we used `updateCount()`
on Gauges when defining the `set()` method for `copyToStatsdAt()`
instead of calling `gauge()`.

The side effect of this is that StatsLib will treat all guages as
counters instead of proper gauge values. This caused an issue with
the `blameStartupRegistry.php` script migration to prometheus format
to break in the way it sends its gauge data to Graphite.

Credits to @Krinkle for the debugging session where we found out that
this was the issue in the library.

Bug: T370636
Co-Authored-by: Krinkle <krinkle@fastmail.com>
Change-Id: I34cf3ce193b354ae732c1d3b0fd3b851d00aeaf0